Refactor abstractions to be internal + use them as delegates
